requirements.txt
flask
python-dotenv
subprocess.run

beautifulsoup4
requests
plan
discord.py[voice]
asyncio